Decouple reporting from server so reporting can be used without running a server on a separate port
