+++ index.js
@@ -1,6 +1,3 @@
-
-
-
 var Metrics = require('./metrics')
   , Reporting = require('./reporting');
 
@@ -10,6 +7,7 @@ exports.Counter = Metrics.Counter;
 exports.Timer = Metrics.Timer;
 
 exports.Server = Reporting.Server;
+exports.Report = Reporting.Report;
 
 exports.version = '0.0.1';
