+++ reporting/index.js
@@ -1,3 +1,2 @@
-
+exports.Report = require('./report');
 exports.Server = require('./server');
-
